Add Initial Repository Documentation
The N-Body simulation code can be written independently of server and client
codebases. While this introduces a bit of overhead in maintaining a separate
library it will allow the simulation code to be consumed in different contexts.
For example, a single server application can consume this library as a
dependency to serve a single simulation to many clients while a bundled client
application can run the simulation locally.

This request adds the [npm][npm] manifest, initial library file, some minimal
documentation.

[npm]: https://www.npmjs.com/