Add Continuous Integration
Problem
=======

The tests for engine are not automatically run.

Solution
========

Add continuous integration with Travis CI. Travis will watch the repository and
run the tests for every push. It then reports the success of the tests in a
multitude of ways. This lets the team know when a new change introduces
regressions or issues.

Howto Test
==========

- [ ] Confirm the tests were run with the PR [in the Travis dashboard][travis_engine]

[travis_engine]: https://travis-ci.org/orbitable/engine/settings